+++ controller/generateInvoice.controller.js
@@ -1,4 +1,4 @@
-const GeneratedInvoice = require("../models/GeneratedInvoice");
+const GeneratedInvoice = require("../models/generatedInvoiceSchema");
 
 // Generate Invoice (associated with an Employee)
 const generateInvoice = async (req, res) => {
@@ -14,8 +14,6 @@ const generateInvoice = async (req, res) => {
       totalGross,
       bankAccount,
       note,
-      createdBy,
-      accountantId,
       banks,
       customerAddress,
       accountantAddress,
@@ -23,6 +21,13 @@ const generateInvoice = async (req, res) => {
       vatRegNo,
       crn,
     } = req.body;
+    let { employeeId, accountantId } = req.user;
+    let createdBy;
+    if (req.user?.isAccountant == true) {
+      createdBy = accountantId;
+    } else {
+      createdBy = employeeId;
+    }
 
     // Create a new generated invoice instance
     const generatedInvoice = new GeneratedInvoice({
@@ -63,9 +68,16 @@ const generateInvoice = async (req, res) => {
 // Update Generated Invoice
 const updateGeneratedInvoice = async (req, res) => {
   try {
-    const invoiceId = req.params.id;
+    let invoiceId = req.params.id;
+    let employeeId = req.user.employeeId;
+    let accountantId = req.user.accountantId;
+    if (req.user?.isAccountant == true) {
+      employeeId = req.user.accountantId;
+    }
     const updateData = req.body;
-
+    const GeneratedInvoiceDetails = await GeneratedInvoice.findById(invoiceId);
+    if (!GeneratedInvoiceDetails.createdBy == employeeId)
+      return res.status(500).json({ message: "not allowed" });
     // Update the generated invoice data based on the provided ID
     const updatedGeneratedInvoice = await GeneratedInvoice.findByIdAndUpdate(
       invoiceId,
@@ -90,14 +102,28 @@ const updateGeneratedInvoice = async (req, res) => {
 const deleteGeneratedInvoice = async (req, res) => {
   try {
     const invoiceId = req.params.id;
+    let { employeeId, accountantId } = req.user;
+    if (req.user?.isAccountant == true) {
+      employeeId == accountantId;
+    }
+
+    //check for valid creator
+    const GeneratedInvoiceDetail = await GeneratedInvoice.findById(invoiceId);
 
     // Delete the generated invoice based on the provided ID
-    const deletedGeneratedInvoice = await GeneratedInvoice.findByIdAndRemove(
-      invoiceId
-    );
-
-    if (!deletedGeneratedInvoice) {
-      return res.status(404).json({ message: "Generated invoice not found." });
+    if (
+      GeneratedInvoiceDetail.createdBy == employeeId ||
+      (GeneratedInvoiceDetail.createdBy == employeeId &&
+        req.user?.isAccountant == true)
+    ) {
+      const deletedGeneratedInvoice = await GeneratedInvoice.findByIdAndRemove(
+        invoiceId
+      );
+      if (!deletedGeneratedInvoice) {
+        return res
+          .status(404)
+          .json({ message: "Generated invoice not found." });
+      }
     }
 
     res
@@ -111,51 +137,96 @@ const deleteGeneratedInvoice = async (req, res) => {
   }
 };
 
-// Get Generated Invoice by Employee ID
-const getGeneratedInvoiceByEmployeeId = async (req, res) => {
+const getGeneratedInvoiceByEmployee = async (req, res) => {
   try {
-    const employeeId = req.params.employee_id;
+    let employeeId;
+    if (req.user?.isAccountant == true) {
+      employeeId = req.user?.accountantId;
+    } else {
+      employeeId = req.user?.employeeId;
+    }
 
-    // Find all generated invoices associated with the provided employee_id
+    const page = parseInt(req.query.page) || 1; // Get the page number from query parameters, default to 1 if not provided
+    const limit = parseInt(req.query.limit) || 20; // Get the limit (number of records per page) from query parameters, default to 10 if not provided
+
+    // Calculate the skip value based on the page number and limit
+    const skip = (page - 1) * limit;
+
+    // Find generated invoices associated with the provided employee_id with pagination
     const generatedInvoices = await GeneratedInvoice.find({
       createdBy: employeeId,
-    });
+    })
+      .skip(skip) // Skip the specified number of records
+      .limit(limit); // Limit the number of records returned
 
     res.status(200).json(generatedInvoices);
   } catch (error) {
     console.error(error);
-    res
-      .status(500)
-      .json({
-        message: "Server error. Could not get generated invoices for employee.",
-      });
+    res.status(500).json({
+      message: "Server error. Could not get generated invoices for employee.",
+    });
+  }
+};
+
+const getAllGeneratedInvoiceByEmployeeId = async (req, res) => {
+  try {
+    let employeeId = req.params.employeeid;
+    let id;
+    if (req.user?.isAccountant == true) {
+      id = req.user.accountantId;
+    } else {
+      id = req.user.employeeId;
+    }
+
+    const page = parseInt(req.query.page) || 1; // Get the page number from query parameters, default to 1 if not provided
+    const limit = parseInt(req.query.limit) || 20; // Get the limit (number of records per page) from query parameters, default to 10 if not provided
+
+    // Calculate the skip value based on the page number and limit
+    const skip = (page - 1) * limit;
+
+    // Find generated invoices associated with the provided employee_id with pagination
+    const generatedInvoices = await GeneratedInvoice.find({
+      createdBy: employeeId,
+    })
+      .skip(skip) // Skip the specified number of records
+      .limit(limit); // Limit the number of records returned
+
+    if (
+      generatedInvoices.createdBy == id ||
+      (generatedInvoices.createdBy == id && req.user?.isAccountant == true)
+    ) {
+      res.status(200).json(generatedInvoices);
+    }
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({
+      message: "Server error. Could not get generated invoices for employee.",
+    });
   }
 };
 
 // Get Generated Invoice by Accountant ID
-const getGeneratedInvoiceByAccountantId = async (req, res) => {
-  try {
-    const accountantId = req.params.accountant_id;
+// const getGeneratedInvoiceByAccountantId = async (req, res) => {
+//   try {
+//     const accountantId = req.params.accountant_id;
 
-    // Find all generated invoices associated with the provided accountant_id
-    const generatedInvoices = await GeneratedInvoice.find({ accountantId });
+//     // Find all generated invoices associated with the provided accountant_id
+//     const generatedInvoices = await GeneratedInvoice.find({ accountantId });
 
-    res.status(200).json(generatedInvoices);
-  } catch (error) {
-    console.error(error);
-    res
-      .status(500)
-      .json({
-        message:
-          "Server error. Could not get generated invoices for accountant.",
-      });
-  }
-};
+//     res.status(200).json(generatedInvoices);
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({
+//       message: "Server error. Could not get generated invoices for accountant.",
+//     });
+//   }
+// };
 
 module.exports = {
   generateInvoice,
   updateGeneratedInvoice,
   deleteGeneratedInvoice,
-  getGeneratedInvoiceByEmployeeId,
-  getGeneratedInvoiceByAccountantId,
+  getGeneratedInvoiceByEmployee,
+  getAllGeneratedInvoiceByEmployeeId,
+  // getGeneratedInvoiceByAccountantId,
 };
